fix: moved load package as well
